move get_settings to tensor_search
